% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.data.bundle}
\alias{as.data.frame.data.bundle.list}
\title{Convert data.bundle's to data.frame objects}
\usage{
\method{as.data.frame}{data.bundle}(x, ...)

\method{as.data.frame}{data.bundle.list}(x, ...)
}
\arguments{
\item{x}{A \code{data.bundle} or \code{data.bundle.list} object.}

\item{...}{Ignored.}
}
\description{
tba
}
\author{
Sven E. Templer
}
