% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Query}
\usage{
get_data(id, limit = NULL, offset = 100)
}
\arguments{
\item{id}{tba}

\item{limit}{tba}

\item{offset}{tba}
}
\description{
This function is not yet implemented as a working method. Stay updated!
}
\author{
Sven E. Templer
}
