% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grps.R
\name{get_grps}
\alias{get_grps}
\title{Query Cologne open data dataset groups of data origin}
\usage{
get_grps()
}
\value{
\code{get_grps} returns a character vector.
}
\description{
tba
}
\author{
Sven E. Templer
}
