% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packages.R
\name{get_packages}
\alias{get_packages}
\alias{get_package_resources}
\alias{get_package_info}
\title{Query Cologne open data dataset packages and resources}
\usage{
get_packages()

get_package_resources()

get_package_info(id)
}
\arguments{
\item{id}{Character string with dataset id as obtained from \code{get_packages}.}
}
\value{
\code{get_packages} returns a character vector.\cr
\code{get_package_resources} returns a list of character vectors.\cr
\code{get_package_info} returns a character vector.\cr
}
\description{
\code{get_packages} queries for a list of available packages (data bundles).\cr
\code{get_package_resources} as above, including ressources (datasets for each package).\cr
\code{get_package_info} queries the information for a package.\cr
}
\seealso{
See \link[CologneOpenDataAPI]{as.data.frame} for how to extract package
metadata as tables.\cr
See \link[CologneOpenDataAPI]{resources} for how to extract package resources
as tables.\cr
}
\author{
Sven E. Templer
}
