% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resources_info.R
\name{get_resources_info}
\alias{get_resources_info}
\title{Query Cologne open data dataset resource information}
\usage{
get_resources_info(id)
}
\arguments{
\item{id}{Character string with the id of the resource.}
}
\value{
\code{get_resources_info} returns a character vector.
}
\description{
Query information of a single resource (dataset).
}
\seealso{
See \link[CologneOpenDataAPI]{get_packages} and \link[CologneOpenDataAPI]{resources}
for how to obtain resource ids.
}
\author{
Sven E. Templer
}
