% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resources}
\alias{resources}
\alias{resources.default}
\alias{resources.data.bundle}
\alias{resources.data.bundle.list}
\title{Extract resources information from package queries}
\usage{
resources(x, ...)

\method{resources}{default}(x, ...)

\method{resources}{data.bundle}(x, simplify = TRUE, ...)

\method{resources}{data.bundle.list}(x, simplify = TRUE, ...)
}
\arguments{
\item{simplify}{Logical value, \code{TRUE} to get result as a data.frame,
with one dataset per row.}

\item{id}{Character string with resource id.}
}
\value{
\code{resources} returns a list of resources lists, or if \code{simplify}
is true, a data.frame.
}
\description{
Get the resources available for datasets stored as \code{data.bundle}.
}
\seealso{
See \link[CologneOpenDataAPI]{get_packages} for how to obtain package resources.
}
\author{
Sven E. Templer
}
